pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
opencv-python = "^4.10.0.84"
numpy = "^2.1.2"
pillow = "^11.0.0"
pytesseract = "^0.3.13"
flask-cors = "^5.0.0"
psycopg2 = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.utils.dev-scripts.run:dev"
